Quasar initialization is now a separate wizard step.
Previously the initialization was run on the EDT in a windowOpened()
callback, blocking the Java user interface. This means that ImageJ did
not respond to e.g. drawing a ROI on an image while the Quasar
initialization was ongoing and the user was not informed of what was
going on.