Change name of Quasar method so it is different from the module name.
We have to avoid Quasar functions that have the same name as the module
they are defined in. This is related to a C# limitation. If they do have
the same name, then Quasar will add a prefix to the function name behind
our backs and the original function will not be defined (but the one
with the prefix will). See email with Bart dd 2017-02-02.