Using the latest anisotropic diffusion code.
Note: the current Quasar release cannot handle (in qlib?) functions that have the same name as the .q file in which they are defined. Otherwise the function will not be defined when the .qlib is loaded. This means that at this point in time anisotropic_diffusion() defined in anisotropic_diffusion.q is not allowed. So we renamed the function to denoise_anisotropic_diffusion() as a workaround.