Removed unused non-local means parameters sigma0 and alpha.
The parameters sigma0 and alpha in DecorrelationParams were actually
parameters for modeling the signal dependency of the noise (not of the
correlation...). However, the signal dependent noise variant of the
non-local means filter is not used anymore since it turned out not to
significantly improve the denoised result, but would force the user to
set even more parameters.