Re-enable the denoise button if the image range is changed.
After denoising a single slice, the denoise button is gone. But if the
user then changes the range to all slices, it re-appears. So, one can
try out the settings on a single slice of a stack, and if the result
looks nice, set the range to all slices and denoise the full stack.