Commented out the decorrelation variant of Non-local Means.
The reason is that the implementation is still incomplete: it uses a
hard-coded, more or less arbitrary, decorrelation kernel. For
decorrelation to be correct the user would need to provide information
on either this kernel itself, or parameters from which we can calculate
the kernel. For now we avoid this more advanced feature and associated
practical difficulties.